Validate ArgGroups after parsing all arguments.
This makes it easier to handle combined switch args (like -abc) and
simplifies the validator at the same time, but not requiring us to
parse the actuall input by the user twice (all Args are already set by
the time validate is called, so we can just check constraints on
them).

To be able to have nice error messages this requires each Arg to track
how it was set on the command line (like using -a vs --all).