Fix bug #25: Remove static variable for ignoring args
https://sourceforge.net/p/tclap/bugs/25/

Move tracking of ignorable args (e.g., after "--") to CmdLine
object. This makes more sense than a static variable and makes 1)
testing easier and 2) it possible to have multiple CmdLine parsing
stages in a single program. There are other static variables that
should be removed too, but that is not easily doable without large
changes to the API (e.g., require every Arg to have a CmdLine object
attached to it) so probably something for 2.0. We could probably
change these to compile time constants in 1.4 to get rid of the
statics.