Enable CMake to build (not test) in 1.2.
This is useful so we can at least check that it compiles on
Windows. We don't bother fixing the tests or testing infra (as has
been done in 1.4 branch) as we expect most development to happen
there. The CMake files will not be included in the distrubution and
are used for development/testing only.