Check arguments vector is not empty before accessing first element.
Some users use TCLAP to parse messages sent between long-running
programs by puting arguments in a std::vector<std::string> and use
CmdLine::parse to parse it. Passing an empty vector to parse results
in a segmentation violation (due to the fact that we always expect to
be able to pop the first element). Change the behaviour to throw a
more user friendly error message.

https://sourceforge.net/p/tclap/bugs/30/